CA-310995: Disable hung task warnings for the idle IO scheduling class
The CFQ IO scheduler has the concept of an idle IO scheduling class
where requests are only scheduled when the disk is idle. However, there
is no mechanism to ensure that these requests complete eventually which
means tasks may be blocked for a long time waiting for a busy, slow disk
to become idle. This is not necessarily a problem but results in scary
warnings in the kernel log that may generate customer support tickets.
Workaround the problem by not printing hung task warnings for tasks with
the idle IO scheduling class and for tasks with the idle scheduling task
(which get scheduled as idle IO by default).

This patch is not for upstream and will not be needed in the future
because the CFQ IO scheduler along with the legacy IO path have been
removed from upstream kernels.