x86/apic: Hide NR_CPUS limit warning
If the host has more CPUs than the kernel is compiled for, warnings
appear in the kernel log. This doesn't make sense for dom0 because the
number of CPUs is set by Xen but it processes the host ACPI tables which
lists all the CPUs. Hide the warning since it is expected.

Example warning:
[ 0.000000] ACPI: NR_CPUS/possible_cpus limit of 24 reached. Processor 27/0x3c ignored.