CA-341597: Increase NR_IRQS
With many VMs, it is possible to hit the IRQ descriptor limit
(IRQ_BITMAP_BITS) since each event channel consumes an IRQ descriptor.
With our configuration:

    IRQ_BITMAP_BITS = NR_IRQS + 8196
    NR_IRQS = 64 * NR_CPUS + 256

Increasing the per-CPU IRQ scaling from 64 to 512 increases
IRQ_BITMAP_BITS from 12548 to 41220 which should be enough to
achieve our current scalability limits.

While it is possible to increase NR_CPUS to achieve a similar affect, this
changes the kABI so it is not an option.